adding first api services

+++ src/main/java/org/example/proyectofinaldwi/repository/EntradaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EntradaRepository extends JpaRepository<Entrada,Integer> {
+public interface EntradaRepository extends JpaRepository<Entrada,Long> {
 }
